useTheming hook
- Inits theming preferences
- handles side effects on html
- provides helper to infer value auto
